Validation functions extracted to a separate file.
